+++ calva/extension.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import * as paredit from "./paredit/extension";
 import * as fmt from "./calva-fmt/ts/extension";
-import * as clojureWarrior from "./clojure-warrior/src/extension";
+import * as highlight from "./highlight/src/extension";
 import * as state from './state';
 import * as jackIn from './nrepl/jack-in';
 import * as util from './utilities'
@@ -71,6 +71,7 @@ function activate(context: vscode.ExtensionContext) {
         fmtExtension = vscode.extensions.getExtension('cospaia.calva-fmt'),
         pareEditExtension = vscode.extensions.getExtension('cospaia.paredit-revived'),
         cwExtension = vscode.extensions.getExtension('tonsky.clojure-warrior'),
+        cwConfig = vscode.workspace.getConfiguration('clojureWarrior'),
         customCljsRepl = state.config().customCljsRepl,
         replConnectSequences = state.config().replConnectSequences,
         BUTTON_GOTO_WIKI = "Open the Wiki",
@@ -219,9 +220,16 @@ function activate(context: vscode.ExtensionContext) {
     state.analytics().logPath("/activated").logEvent("LifeCycle", "Activated").send();
 
     if (!cwExtension) {
-        clojureWarrior.activate(context);
+        highlight.activate(context);
     } else {
         vscode.window.showErrorMessage("Clojure Warrior extension detected. Please uninstall it before continuing to use Calva.", ...["Got it.","Will do!"]);
+    }
+
+    for (const config of ["enableBracketColors", "bracketColors", "cycleBracketColors", "misplacedBracketStyle", "matchedBracketStyle", "commentFormStyle", "ignoredFormStyle"]) {
+        if (cwConfig.get(config) !== undefined) {
+            vscode.window.showWarningMessage("Legacy Clojure Warrior settings detected. These settings have changed prefix/namespace to from `clojureWarrior´ to `calva.highlight`. You should update `settings.json`.", ...["Roger that!"]);
+            break;
+        }
     }
 
     return {